Move logger to app folder
