+++ app/models/invite.server.ts
@@ -1,7 +1,7 @@
 import type { Invite } from "@prisma/client";
 
 import { prisma } from "../db.server";
-import { logError } from "../utils/logger.server";
+import { logError } from "../logger.server";
 
 export async function redeemInviteCode(inviteCode: Invite["id"]) {
   const existingInvite = await prisma.invite.findUnique({
