+++ app/session.server.ts
@@ -3,7 +3,7 @@ import invariant from "tiny-invariant";
 
 import type { User } from "./models/user.server";
 import { getUserById } from "./models/user.server";
-import { logError } from "./utils/logger.server";
+import { logError } from "./logger.server";
 
 invariant(process.env.SESSION_SECRET, "SESSION_SECRET must be set");
